requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
discord.py==2.3.2
schedule==1.2.1
lxml==5.1.0
Pillow==10.2.0
openai==1.12.0